🛠 rename: Modify AlertDialog.tsx component renaming and import
